Dataplane refactoring. Rebase #816 to latest
